day 39
my body is a machine that turns fuel into Cod4 Decomp code. My fuel is a lean dual-mixture of my hatred for the currently popular FPS titles, mixed 50% for my past admiration and nostalgia for the title.

+++ KisakCOD/src/universal/com_math_anglevectors.cpp
@@ -78,7 +78,7 @@ void __cdecl AnglesToAxis(const float *angles, float (*axis)[3])
     (*axis)[8] = cr * cp;
 }
 
-double __cdecl Vec4Normalize(float *v)
+float __cdecl Vec4Normalize(float *v)
 {
     float v2; // [esp+0h] [ebp-Ch]
     float ilength; // [esp+4h] [ebp-8h]
